Add AnnotationGuide
