+++ alembic/versions/19687a7ba1e1_add_annotationguide.py
@@ -1,0 +1,41 @@
+"""Add AnnotationGuide
+
+Revision ID: 19687a7ba1e1
+Revises: 9d772a357dbc
+Create Date: 2020-05-11 08:37:43.344538
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '19687a7ba1e1'
+down_revision = '9d772a357dbc'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'annotation_guide',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('label', sa.String(), nullable=False),
+        sa.Column('url', sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint('id', name=op.f('pk_annotation_guide'))
+    )
+    with op.batch_alter_table('annotation_guide', schema=None) as batch_op:
+        batch_op.create_index(batch_op.f('ix_annotation_guide_label'), [
+                              'label'], unique=False)
+
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('annotation_guide', schema=None) as batch_op:
+        batch_op.drop_index(batch_op.f('ix_annotation_guide_label'))
+
+    op.drop_table('annotation_guide')
+    # ### end Alembic commands ###
