--- alembic/versions/3bc14a7bc467_add_label_model.py
@@ -1,42 +1,0 @@
-"""Add Label model
-
-Revision ID: 3bc14a7bc467
-Revises: 9d772a357dbc
-Create Date: 2020-05-11 08:32:47.813270
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '3bc14a7bc467'
-down_revision = '9d772a357dbc'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        'label',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('name', sa.String(), nullable=False),
-        sa.Column('annotation_guide_url',
-                  sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint('id', name=op.f('pk_label'))
-    )
-    with op.batch_alter_table('label', schema=None) as batch_op:
-        batch_op.create_index(batch_op.f('ix_label_name'), [
-                              'name'], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('label', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_label_name'))
-
-    op.drop_table('label')
-    # ### end Alembic commands ###
